finalbg-main/requirements.txt
fastapi
uvicorn[standard]
celery
redis
requests
sentry-sdk[fastapi]
deep-translator
appwrite
minio
qdrant-client
python-dotenv
numpy
opencv-python
scikit-learn
Pillow
sentence-transformers
transformers
slowapi
pydantic-settings
